custom-elements/instantiating-custom-elements/custom-element-type-local-name-test-002.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element type is the local name of the custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="The custom element type can be given to a custom element at the time of its instantation as the local name of the custom element">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

function doTest(element) {
    var doc = newHTMLDocument();
    
    var name = 'x-foo-' + element;
    
    var obj = doc.createElement(element);
    
    var XFoo = doc.register(name, {
        extends: element
    });
    
    doc.body.innerHTML = '<' + name + ' id="customElement"></' + name + '>';
    
    var xfoo = doc.querySelector('#customElement');
    
    assert_true(xfoo instanceof XFoo, 'Wrong instance type');
    
    assert_class_string(xfoo, obj.constructor.name,
            'Custom element  must be a ' + obj.constructor.name);
}


var parameters = [];

for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
    parameters[i] = ['Custom element type ' + HTML5_ELEMENTS[i], HTML5_ELEMENTS[i]];
}

generate_tests(doTest, parameters, 
        'Custom element type is the local name of the custom element. '
        + 'Element is created in HTML');

</script>
</body>
</html>
